Refactored files to be simpler subcomponents.  Start game button works to start and pause game. Reset Timer added to reset the timer to 15:00.
