+++ src/App.css
@@ -109,10 +109,11 @@
 }
 
 .buttons {
-  width: 50%;
+  width: 65%;
   margin: 0 auto;
   display: flex;
   justify-content: space-around;
+  display: flex;
 }
 
 .homeButtons,
@@ -122,6 +123,7 @@
   justify-content: space-around;
   align-items: center;
   width: 100%;
+  height: 20%;
 }
 
 .homeButtons__touchdown,
@@ -129,7 +131,9 @@
 .awayButtons__touchdown,
 .awayButtons__fieldGoal,
 .downButton,
-.quarterButton {
+.quarterButton,
+.startButton,
+.resetButton {
   margin: 1rem;
   width: 175px;
   border-radius: 4px;
